misc/whysapp/requirements.txt
plyplus==0.7.5
pycryptodome==3.8.1

